// src/app/models/eleve.ts
import {Genre} from "./genre";

export interface Eleve {
  id: null | number
  prenom: "string"
  nom: "string"
  adresse: "string"
  genre_id: number
  date_naissance: Date
  annee:number
  prenomParent: string
  nomParent: string
  adresseParent: string
  telephone: string
  cni: string
  login: string
  genreIdParent: number
}